Fixed layout on search page and profile page
Put Anime title above picture on search results. Added a scrollable friends list to the right side of the profile page that properly adjusts to the different screen sizes. Added multiple breakpoints in the css to adjust all of the elements on the profile page as the screen scales up and down.